Rename game info config better
