+++ src/config/gameInfo.json
